+++ src/util/Utils.h
@@ -14,7 +14,7 @@ public:
 
     static std::vector<std::vector<int>> readVList(const std::string &file_path);
 
-    static float stdDev(std::vector<float> &items, float average);
+    static float stdDev(std::vector<int> &items, float average);
 
     static void floydWarshall(std::vector<std::vector<float>>& G);
